requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11
matplotlib>=3.7
Pillow>=9.5
PyYAML>=6.0
statsmodels>=0.14
datasets>=2.14
